Streamline Format functions tests
In having separate test suites for isMarkdown and not, each builder had
its own coverage gaps. The two MessageBuilders should share almost all
tests.
