+++ formatter/src/commonTest/kotlin/FormatBodyTest.kt
@@ -4,11 +4,23 @@
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */        
+ */
 
+import info.DEFAULT_GIT_COMMENT_CHAR
 import kotlin.test.Test
 
-class FormatBodyTest {
+abstract class FormatBodyTest {
+
+    abstract val isMarkdown: Boolean
+
+    protected fun formatBody(
+        message: String,
+        commentChar: Char = DEFAULT_GIT_COMMENT_CHAR,
+    ) = formatBody(
+        message,
+        commentChar,
+        isMarkdown,
+    )
 
     @Test
     fun reformatsBodyGivenBodyLineOver72() {
@@ -49,29 +61,5 @@ class FormatBodyTest {
             "01234567890123456789012345678901234567890123456789012345678901234567890"
         ) shouldEqual
             "01234567890123456789012345678901234567890123456789012345678901234567890"
-    }
-
-    @Test
-    fun whenFormatBodyWithMarkdownOptionFalseThenFormatsAsPlainText() {
-        formatBody(
-            """
-                # H1
-
-                01234567890123456789012345678901234567890123456789012345678901234567890
-                foo foo foo foo foo foo foo foo foo foo foo foo foo foo foo foo foo foo
-
-                ```kotlin
-                println("snippet")
-                println("snippet")
-                ```
-            """.trimIndent()
-        ).shouldEqual(
-            """
-                01234567890123456789012345678901234567890123456789012345678901234567890
-                foo foo foo foo foo foo foo foo foo foo foo foo foo foo foo foo foo foo
-
-                ```kotlin println("snippet") println("snippet") ```
-            """.trimIndent()
-        )
     }
 }
